\name{odfWeave.survey-package}
\Rdversion{1.1}
\alias{odfWeave.survey-package}
\alias{odfWeave.survey}
\docType{package}
\title{Support for odfWeave on survey objects}
\description{
Provides methods for \code{odfTable} for \code{svytable}, \code{svystat}, and \code{svrepstat} objects from the survey package.
}
\details{
\tabular{ll}{
Package: \tab odfWeave.survey\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-14\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Thomas Lumley <tlumley@u.washington.edu>
}
\keyword{ package }
